+++ include/paper_clock.h
@@ -1,0 +1,7 @@
+#pragma once
+#include <WebServer.h>
+namespace PaperClock {
+void begin();
+bool ready();
+void routes(WebServer& server,const String& token);
+}